emul_tps6699x: Add emulator support for set drp mode
Add set drp mode and enable test. TI driver doesn't
support TRY_SNK. Created an emulator API to fetch
supported DRP modes per driver. If the emulator does
not support the API the defaults are used as defined
in the test. This will allow reuse of this test for
future PDC drivers.

BUG=b:339686045, b:345292002
TEST=./twister -T zephyr/test/pdc

Change-Id: Ia2babaf13958ff4f243faf51d2cd486f872aeec1