zephyr: console: Don't filter help output
The shim that connects EC legacy commands to the Zephyr shell used
printk() directly.  This caused legacy command help to be filtered if
the CC_ZEPHYR_LOG channel was turned off.

Send legacy command help directly to the Zephyr shell, which avoids the
filtering done by the printk hook.

The new unit test fails when the help text uses print() and passes when
using cprintf().

BUG=b:359744043
TEST=Run "chan 0" and "pd 2 state" on tentacruel.  Verify help text
shown.
TEST=./twister -ivc -T zephyr/test/console

Change-Id: Ice9e26ded1dd418311e18ad6bc8ce4b90dbc0f5f