EC/ISH: Add environment variable to find protoc
Pigweed looks for protoc under the CIPD_INSTALL_DIR environment
variable. While there's nothing saying that it actually has to be
installed from CIPD, we're stuck with this name until the chroot
migrates to Python 3.11 and we can un-pin Pigweed. For now, we use
this variable to tell Pigweed where protoc is installed.

We also require PW_ROOT and NANOPB_DIR which help us tie all the pieces
together.

BUG=b:345530593
TEST=Build brox-ish with a PW proto library

Change-Id: Ib971ed0020f3f650af5a28d5b5d0485331f318bc