EC/ISH: Add NAMED_I2C_PORT_COUNT
We're currently required to have the named i2c ports, but tests that
don't use the ports can be broken by having them in there. Removing
all named ports breaks the build because the enum is empty. With this
addition, we have have 0 named i2c ports in a build.

BUG=b:345530593
TEST=Build brox and unit tests.

Change-Id: Ib04271644ec2c62e2626092cbc7d11d829ffb345