Test with max size battery names
The max size of a battery manufacturer or model is 31 characters (not
32), because the block contains 1 byte of length, and 31 bytes of data.
Set the emultated battery in native_sim tests to the largest possible
size.

Fix the battery_v2 test to ignore the trailing nul character, so that it
passes when the name is > 7 chars (for v0) and >11 chars (for v1). I
really don't know why the host command returns null terminiated strings
in fixed sized buffers, but that consumes a byte.

BUG=b:362163827
TEST=./twister

Change-Id: I87630af45414270a92b877bce74b8a920317c4a0