subsys: usb_dc: add usb state printout and set hid log level to warning
The USB state is currently logged by the Zephyr HID framework.
However, the developer cannot check the USB state from the EC console
when the firmware is running in the RO area, as the HID function is
disabled in the RO image. This change introduces a USB state printout
in the application layer and adjusts the HID log level to 'warning.'

BUG=none
TEST=none
LOW_COVERAGE_REASON=keyboard and touchpad test for Zephyr still not implement

Change-Id: Id24691168813dab770dd9018a13f500667bcf2a3