host_sleep: mutex for sleep_signal_timeout
The sleep_signal_timeout needs to add the mutex to avoid the use of
sleep_complete_resume() and sleep_resume_transition() at the same time.

If call the sleep_signal_timeout at the same time. It will cause the
resume from sleep. Then it will do the sleep hang process.

BUG=b:358516419
TEST=zmake build SUCCESS

Change-Id: Ie4896a599dbf9d9ac05794255e5d39607ab575aa