brox: fix AP power sequence when logging is off
When logging is off, and the ITE chipset is configured for running from
RAM, the AP power sequence advances to the S3S0 state before the
ALL_SYS_PWRGD signal is ready.

Enable interrupts on this signal so that the AP power sequence thread
restarts after ALL_SYS_PWRGD comes ready.

BUG=b:362675141
TEST=Run "chan 0" and "powerbtn" on the EC console. Verify Brox boots to
S0.

Change-Id: I272a33bbdf457a1114795f5f9c6abae0e476c5fc