ec: Align stdlib assert with builtin ASSERT macro
The assert function within the stdlib ends up calling _exit, which sets
PANIC_SW_EXIT. Modify it, such that it calls PANIC_SW_ASSERT, when there
is an assertion failure. Align the info of the C standard library to
that of the builtin ASSERT macro. With that change, we can align the
fail_regex within the run_device_tests.py, now the stderr messages are
aligned.

BUG=b:368324905,b:361641154
TEST=run_device_tests.py -b bloonchipper -t assert_stdlib

Change-Id: I44f2d8b9a47d8249ef590cacb14c771208c0380b