ectool: pdchipinfo: Restrict version to highest supported by ectool
The `ectool pdchipinfo` implementation queries the EC for the highest
version of EC_CMD_PD_CHIP_INFO supported and uses that version of the
host command, but neglects to consider if ectool itself can handle that
version. This means a newer EC compared to ectool can encounter a
"message too long" error.

Add a check to clamp the command version to the highest that ectool
currently supports, which is affected by the response struct version
used and the decoding logic.

BUG=b:364408259
TEST=Change max supported version to 2, tets against EC with maxver of 3

Change-Id: Ib4692b0f2d130020a4d135dadfe23668d623af72