orisa-ish: Support to notify EC for clamshell or tablet mode change
In orisa, sensors are supported by ISH and tablet mode change is
updated to ISH by motionsensor task, not to EC. EC also need to be
updated for mode switch as it need to update AP for mkbp switches.
Hence ISH need to notify EC for clamshell or tablet mode change.

gpio_soc_ec_ish_nb_mode_l is an active low pin and It is configured
as output pin from ISH to EC. ISH will notify the EC about the clamshell
mode or tablet mode change by updating this pin.

BUG=b:347811875
TEST=zmake build orisa-ish

Change-Id: Id319dd8a8bd396be23e1e69f67c6bd7717d5d349