pdc_power_mgmt: Synchronize sink path control
Enabling sink path on both ports may result in reverse current to one of
the port partners. When switching the active charge port, the existing
sink path should be turned off before turning on the new one. When the
system is not powered by the battery, switching active charge port
shouldn’t be allowed.

Currently, the pdc power management threads run independently for each
port. The order of tuning on/off sink path is not guaranteed. To ensure
the sink path control synchronization, the state machine is modified so
that the sink path is not enabled by default.
pdc_power_mgmt_set_active_charge_port() controls the order of turning
on/off the sink paths.

BUG=b:354836076
TEST=Attach two type-c chargers, two PD chargers or one type-c and one
PD charger to the DUT and probe the sink path signal to make sure the
sink path is not enabled simultaneously when charging power switching
occurs. The same test is repeated for EC reboot and no battery
condition.

Change-Id: Ia351fe0c3d44cfe19b1a9d20edf3a3f74b5a2d97