ptlrvp: Gpio Pin Indication to Soc for CSE early RO RW jump
For boot time optimization, CSE is initially looking for a GPIO
indication from soc side, to make decision on, whether to do an early
RO to RW jump in the early boot stage, without any global reset. In the
case of recovery request, it must continue to remain in RO. For this
purpose, EC handles the GPIO to soc side, based on whether boot request
is for Normal boot or for Recovery request. In the case of normal boot,
EC deassets the GPIO pin, while during recovery boot, EC asserts it.

BUG=b:299055472,b:259659012
LOW_COVERAGE_REASON=Coverage to be handled in b/336751447
TEST=Tested on PTL RVP

In normal boot, CSE RO to RW jump occurs without global reset, in early
boot.

Change-Id: I16357541bf638866ceec65b8cc6da40807ae6be9
Signed-off-by: Poornima Tom <poornima.tom@intel.com>