ptlrvp: Gpio Pin Indication to Soc for CSE early RO RW jump
This is part of boot time optimization in CSE. CSE is initially looking
for a GPIO indication at soc from EC, to decide for early RO RW jump,
based on if the boot request is because of recovery or not. More details
in cross-bug-b:259659012.

BUG=b:299055472,b:259659012
LOW_COVERAGE_REASON=Coverage to be handled in b/336751447
Test=Tested on PTL RVP

In normal boot, CSE RO to RW jump occurs without global reset.

Change-Id: I16357541bf638866ceec65b8cc6da40807ae6be9
Signed-off-by: Poornima Tom <poornima.tom@intel.com>