chip/stm32: Larger console UART buffer for STM32L5
The STM32L5 chip has 256kB memory, so we can affort more than the
default 512 bytes of console UART buffer.

This is particularly useful when debugging by attaching to the UART
console, while a host computer is sending a sequence of commands via the
USB console channel.  The default small buffer would quickly overrun,
causing the UART transcript to be incomplete.

BUG=none
TEST=make build BOARD=hyperdebug  (the only board using this chip)

Change-Id: Id56c0292fee87b6ec9eed920f82facaca4c366e6