gsctool: Add GetChipId TPMV command
Add processing for GetChipId TPMV command for automatic GSC device
selection. If this TPMV command exists, then we use it to determine the
GSC device, otherwise we fall back to using the tens digit of the major
version component.

See CL:*7675534 for Ti50 implementation of TPMV command.

BUG=b:364705511
TEST=verify with chip that doesn't implement TPMV command that version
is still used as a fallback
TEST=verify with chip that does implement TPMV command that TPM VID_DID
is used for selection instead of version.

Change-Id: I810e0b4f03b130420eecc91be7498cacebcc142c