Import tpm_generated and its dependencies.
This copies over the trunks TPM command generation code (tpm_generated.cc), as well as its unit test and dependencies. It also copies over secure_hash from libchrome, as libchrome-crypto is a device-only library and we want to be able to test code that depends on this library on the host.

These files are copied over as-is from the `main` branch; there are no local changes. The exception is `Android.bp`, which I created with a comment explaining the purpose of this directory.

BUG=b:359274525

Change-Id: I9b125f38899c45594dd6b03f340e9f00c4bdfb42