TCPMv2: stop drawing current for 5V@0A contract
Per PD r3.1 v1.8 ss 6.4.1.2.2, when a source wants a sink,
consuming power from Vbus, to go to its lowest power state,
the voltage[bit 19:10] shall be set to 5V and the maximum
current shall be set to 0mA. This is used in cases where
the source wants the sink to draw pSnkSusp (25mW).

So I add sink_stop_drawing_current() for 5V@0A contract.

BUG=none
TEST=pass TEST.PD.PS.SNK.1#11 on rauru

Change-Id: I706e5ed0064dafb0c762fe33bad549478b2922fb