test: fix test flake on test_new_pd_sink_contract
test_new_pd_sink_contract failed to set any PDOs for the attached source
partner. This could cause the PDC thread to disable charging on the
port, in which case it is expected that the sink path would be disabled.

Explicitly set partner PDOs prior to the connection and before
triggering the battery charging status change.

BUG=none
TEST=./twister -ivc -T zephyr/test/pdc

Change-Id: I9b60f058c470df81624ee7177c3fca5bb75006b9