test: Support buccaneer power utilization testing
We left the default of "ppvar_fp_mw" for other boards intact
We don't use the ppvar_fp, because that only applies to the FPC
and is actually incorrect for the FPC sensor, too. See b/368415325.

The Elan 80SG functions completely on the 3.3V rail. There are some
signal buffers that use ppvar_mcu, but these should be considered
more losses, than performance figures. Plus, these buffers probably
don't exist on Chromebook.

Given these power figures and the fact that the values shown
in "dut-control  -- -t 10 pp3300_fp_ma" seem to be fluctuating
quite a bit, I'd say that the sensor is not in power down mode.

BUG=b:339491623
TEST=./test/run_device_tests.py --board=buccaneer --with_private=no \
       --flasher=servo_micro --test power_utilization_{idle,sleep}

Change-Id: Iab878f8cf4d5075bf2ee40ebd4b34de967bc905a