ec: Add min-pagesize to avoid warning
GCC 11 and above may generate a warning when dereferencing a constant
address pointer whose address is below the value specified by the
`min-pagesize` parameter (defaults to 0x1000). The `min-pagesize` parameter
is set to 0 such that GCC never generates any warnings for the constant
address pointers. For more details, refer to the GCC PR99578.

BUG=b:367332648
TEST=./firmware_builder.py

Cq-Depend: chromium:5949063
Change-Id: I2c2ec16de83872d127779bc5cc4be6a34f8a5789