itecomdbgr: add a retry during write enable
The first write enable will often have a timeout in enabling the write
enable latch.  Add a retry as a workaround until the root cause is
identified.

This change also fixes reads of the entire image. De-assert the
chip-select after a read command before writing the data out to the
filesystem. This also drops an unneeded read to the status register
after each page read.

BUG=b:368601076
TEST=Pass UART based programming on Brox using CCD 10 consecutive times

Change-Id: I0bf6260e89da0960f2cd2ac4e87930f3b5bca848