gsctool: Handle special case for version 0.21.x
Version 0.21.x does not respond with an error code if an unknown TPMV
command is requested; it just never respond which ultimately leads to a
USB timeout on the host side. To prevent this, care out version 0.21.x
as a DT specific version without calling a TPMV command that is does
not know about.

BUG=b:368631328
TEST=Resigned the original 0.21.1 release for my devboard and veriied
     that this CL indeed fixes firmware update.

Change-Id: I0f36a82351ea125e58410457d5702ce6b584150b