test/run_device_tests.py: Work around hang on exit
The run_device_tests.py always hangs after running the tests, which
prevents us from running the script with automated tools that want to
check the exit code.

The hang appears to be an issue with the ThreadPoolExecutor attempting
to wait for pending work. For now we'll force an exit with os._exit
instead of sys.exit.

BUG=b:368684364
TEST=./test/run_device_tests.py --board bloonchipper --renode -t aes

Change-Id: I89e1629f66f05390d8b23b430e956f2c1a7b9b00