Remove shurst@google.com from board/servo_v4p1/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/shurst@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I477630bc5dc5c629d78adf245c817e00f43b94da