zephyr: fpmcu: add restricted console unit test for fpupload command
Add a unit test for the fpupload command when (a) the system in unlocked
and (b) the system is locked. This is done in fingerprint/task in order
not to duplicate the board overlay_sim file and the configurations. We
can split the fpsensor_debug tests to a separate source file.

BUG=b:368109581
TEST=./test/run_device_tests.py -b bloonchipper -t restricted_console
     --zephyr && ./twister -v -i -T zephyr/test/fingerprint/task

Change-Id: I52e9b144580742f9a07ff4f19a99168a4854ccdf