ec: Fix compare_builds.sh
Move toolchain_init to a separate python script to be imported by the
firmware_builder scripts.  Make the new script individually callable
and have it return a formatted string which can be parsed as a bash
associative array.  Update compare_builds.sh to set the environment
by calling init_toolchain and writing the resultant environment into
the temp Makefile used by the script to compare builds.

BUG=b:370803752
TEST=./util/compare_build.sh -b all -j 160

Change-Id: Ibadf8ea2d026997e9f47a7edcbf8b23b36bd74bb