pdc_rts54xx: Remove notification on USB PD Attention workaround
The rts54xx previously did not support enabling notifications on USB PD
Attention messages. To enable this feature the rts54xx driver manually
sets the attention bit of the connector status change bits when the
driver suspect an IRQ came from a USB PD Attention message (DP Status).

Now that the rts54xx correctly supports sending an IRQ on USB PD
Attention, remove the workaround to signal an Attention message from the
driver.

BUG=b:370042671
TEST=zmake build brox, ./twister -T zephyr/test/pdc, manually checked
that the PDC sets the attention connector status change bit when sending
an IRQ in response to a USB PD attention.

Change-Id: I436a07789d488967bc3d104622a5df0a510a1375