keyboard: clean up some Kconfig options
The dep from PM_DEVICE_RUNTIME was incorrectly set on the col2 driver,
should have been on the whole keyboard one. Move it to the right
location, and also convert it to a "select" so it does not have to be
picked manually, and switch the INPUT to a select as well so that once
the devicetree label is set correctly there's less noise to deal with in
the options.

BUG=none
TEST=zmake compare-builds -a

Change-Id: I6b52aa4385e7adec965fe5bad032a5ec811fa1c4