bujia: ops: Implementation SYS_FAN controlling
SYS_FAN  is connected to EC from the OPS (Open Pluggable
Specification) JAE connector. This signal allows system to control
the FAN at OPS partner.

To get better performance, we can enable SYS_FAN if the system FAN
is spinning up.

BUG=b:327978076
TEST=On Bujia. SYS_FAN is disabled on suspend/shutdown.
SYS_FAN is enabled on S0 if the system FAN is spinning up.
SYS_FAN is disabled on S0 if the system FAN is not spinning up.

Change-Id: I36ecfcd6a59035fc090d9bdaedd43b7f4aca505d