test: Enable panic_data even when no RO specified
This enables the panic_data test for all local build, even
before we have locked down the RO. This is useful for shaking out
issues with panic_data during normal bringup of a new platfom, rather
after locking down the RO image.

I considered gating the addition of this test on when no variants
were specified, but then realized that adding a new helipilot follower
would then not trigger this test. I then considered scanning all
variants for an RO firmware that matched the board name to trigger
adding this test, but then that would omit testing when bringing
up a new RO variant. Let's just keep this simple and add the test
for all cases.

BUG=b:336640650, b:336640151
TEST=./test/run_device_tests.py --board=helipilot --with_private=no \
       --flasher=servo_micro -t panic_data

Change-Id: I86516b24be14b579f9e13798d2e963b61da6e253