test: Fix panic_data test for npcx
On stm32 parts of the panic frame are preserved in the backup ram/regs
during a hard reset. To save space values saved in these regs are 16
bit values. For Nuvoton chips, the panic frame is simply persisted in
data RAM, with all original panic components. There is only one npcx
board that seems to persist panic info in backup ram, but even then,
it doesn't appear to be truncating the value to 16bit.

Although stm32 was implemented first, it appears that the 16bit
truncation of panic data is now the less common approach, thus we
single out stm32 as the odd case in this test.

BUG=b:336640650, b:336640151
TEST=./test/run_device_tests.py --board=helipilot --with_private=no \
       --flasher=servo_micro -t panic_data

Change-Id: I9e846941758e5e2543dee91a311659c622d53b93