Remove libchrome dependencies from tpm_generated.
This removes all dependencies tpm_generated has on libchrome. Where practical, the code was migrated to alternatives available in AOSP. For a few things that are used hundreds or thousands of times, this either uses a `#define` to replace the libchrome use with Android equivalents or adds a minimal implementation of the libchrome functionality.

Change-Id: Ib59a0024fe2df8c075a83d058986749224c50ee1