pdc: Allow overriding PDC_SM_SETTLED_TIMEOUT_MS
Allow boards / unit tests to choose a different value for
PDC_SM_SETTLED_TIMEOUT_MS, the timeout for waiting the on the PDC power
mgmt state machine to settle on a steady state. By default, this value
is pegged to the PDC command completion timeout of 2000ms and remains
unchanged.

BUG=b:372259212
TEST=zmake compare-builds brox

Change-Id: I044969875b607116e2f8dbc796df4caf5a57a626