docs: Fix race condition in fp factory requirements
The star example of this is when capturing the pattern0/1 samples,
where the MKBP event may be emitted before the user/script starts
waiting for the event using "waitevent FINGERPRINT 500". Elan's
capture for pattern1 is very fast and usually triggers this race
condition.

This also makes the human fp capture robust against this capture
race condition, between waiting for events and touching the sensor.
That means that we could now use this capture command in automated
scripts, where is possible that the finger press might occur
immediately after starting the capture.

BUG=b:372316042
TEST=Test on buccaneer

Change-Id: I2e2d3e08ad537e4a72ab2835c890b11f83d580dd