pdc: Use the sink path status to set initial charge port
Use the sink path status from the connector status to set the intial
charging port.  If the system doesn't have a battery, this ensures that
the charge manager knows which port is providing power and the charge
manager avoid changing the source.

Note that only one test combination is failing with these changes. With
the battery disconnected, connecting a suzyQ on port 0 followed by a 15W
non-PD charger on port 1, the sink enable signals asssert simutaneously
and at the wrong voltage.

BUG=b:354836076
TEST=See b/354836076#comment71 for a full list of tests run.
TEST=./twister -ivc -T zephyr/test/pdc

Change-Id: Ib525e8b14125e0dc24d5b5c4e852b02dc3a94431