common: Enable SRAM MPU protection on helipilot
Protects the code RAM section of the helipilot board, which executes
code from RAM. This mechanism protects the code from being modified using
the MPU. The MPU protection are setup at boot using post_mpu_init block.
This prevents write access to code RAM after system_is_locked is
detected and flash is initialized.

The CONFIG_RAM_LOCKED is moved to be defined only when SECTION_IS_RW is
defined because there is no mechanism to disable it before jumping from
RO to RW.

This was the previous landing of SRAM MPU protection
(http://crrev.com/5879391). The mechanism to lock the system was
modified in this CL, where it is currently using hardware right protect
implemented in  run_test_devices.py. In the previous attempt, software
right protect approach was used by forcing the system_is_locked function
to always return 1. This was interfering with the ability of the system
to jump from RO to RW during the test execution.

BUG=b:3246677118
TEST=run_device_tests.py -b helipilot -t sram_mpu_protection

Change-Id: I72ee1622b7f477411625aaf897bcf4a4cc7a5718