TCPMv2: Send full-length Status by default
Update the logic for responding to Get_Status based on the partner's
reported PD revision. If a partner does not respond affirmatively to
Get_Revision, thereby leaving its supported PD revision ambiguous,
respond to Get_Status with the 7-byte Status Data Block defined starting
in PD r3.1. If a partner does respond affirmatively with Revision,
maintain the distinction between a 6-byte SDB in PD 3.0 and a 7-byte SDB
in newer revisions.

BUG=b:366241394
TEST=twister -s zephyr/test/drivers/drivers.default

Change-Id: I4b19830f5366dbf0c0d8ff2fd4df88ac3614fbd5