deku: Enable POWER_BUTTON_INIT_IDLE config option
This CL enables CONFIG_PLATFORM_EC_POWER_BUTTON_INIT_IDLE which is
used for Chromebox designs to allow the AP to remain in S5 with an AC
connection following a graceful shutdown.

BUG=b:182876692
TEST=Verifed on Deku that following shutting down the AP via the power
button, then removing AC power, when AC power is reconnected that the
AP doesn't boot until the power button is pressed. Without this config
option, the AP would boot up in the same scenario.

Change-Id: I6ca33cc16d305bbe8c010bec6491226f6208c917