test: Create sram_mpu_protection test
Create a test to verify SRAM MPU protection and confirm that the system
jumped to RW. This was the previous landing of this test
(https://crrev.com/c/5879391). The testing mechanism was modified in
this CL:

(1) The locking of the system is achieved by a combination of
hardware / software write protect mechanisms implemented in
run_device_tests.py. In the previous attempt, software write protect was
achieved by forcing system_is_locked function to always return 1. This
was interfering with the ability of the system to jump from RO to RW
during the test execution (b/374822463).

(2) The unit tests currently runs when SECTION_IS_RW is defined to
ensure that the system has jumped to RW.

BUG=b:324677118
TEST=run_device_tests.py -b helipilot -t sram_mpu_protection.*

Change-Id: I72ee1622b7f477411625aaf897bcf4a4cc7a5718