pdc: ti: Upgrade TI base FW to 19.9.2 and appconfig JSON
Upgrade the bundled TI base FW to version 19.9.2.

Import the existing JSON appconfig (GOOG0270) and re-export it to
upgrade the appconfig data to the latest format. Theoretically this
should not change the functional configuration, but new fields have been
added and configured to TI's defaults.

Disable USB4 and enable FRS in config.

BUG=b:373953632, b:376498698
TEST=Flash to Orisa DUT

Change-Id: I7b4103d7ecef74a0480dd0143aa8f481148e4421