keyboard_scan: only do the extra col2 delay when changing col2
The col2 line requires an extra 30us delay when being asserted or
deasserted, right now this is added to the normal delay by changing the
board configuration and it's applied to all lines, but that is only
required when asserting or deasserting col2, leading to some unnecessary
delays when scanning the other columns.

Change the scanning logic to only delay when selecting or deselcting
col2, which saves few cpu cycles that are otherwise spent spinning.

This requried tweaking all board configs that were overriding the delay,
some of those are now back on the default config from keyboard_scan.c.

Drop few overridden structures that are now redundant, and few comments
that makes it looks like they are redundant when they are not.

BUG=none
TEST=build and run on rex, check all the keys with ksstate on, no false
triggers

Change-Id: Ic603760054b10a16a69c6b839da54a77655f7909