pdc: Remove policy to downgrade source caps based on RDO
The policy to downgrade an attached sink back to 1.5A if the device RDO
set the operating current to 1.5A or less was originally added to comply
with the "Host Source Power Provising" test.  This test has been deleted
from the PD 3.1 compliance suite.

The downgrade policy was causing problems with TEST.PD.PS.SRC.01, so
remove the downgrade because of RDO code.

This change only fixes PDC based implementations.  We likely need a
similar change for TCPMv2.  See b:361645297, b:312258082, b:239140539

BUG=b:369715317
TEST=Run TEST.PD.SRC.01 against Lotso

Change-Id: I8767ea7d636369740cbc02a6404ee4aa4547d3ea