pdc: Remove policy to downgrade source caps based on RDO
The policy to downgrade an attached sink back to 1.5A if the device RDO
set the operating current to 1.5A or less was originally added to comply
with the "Host Source Power Provising" test.  This test has been deleted
from the PD 3.1 compliance suite.

The downgrade policy was causing problems with the following tests in
the PD 3.1 CTS:
  TEST.PD.PS.SRC.1
  TEST.PD.PS.SNK.3
  COMMON.CHECK.PD.14 on TEST.PD.SRC.1_Rev2Src
  COMMON.CHECK.PD.5 on TEST.PD.SRC.1_Rev2Src

This change only fixes PDC based implementations.  We likely need a
similar change for TCPMv2.  See b:361645297, b:312258082, b:239140539

BUG=b:369715317, b:375662743, b:375668955, b:375668966
TEST=Run TEST.PD.SRC tests gainst Brox

Change-Id: I8767ea7d636369740cbc02a6404ee4aa4547d3ea