extra/um_ppm: Fix usermode portion of um_ppm
um_ppm is no longer functional as its functionality was incorporated
into Zephyr. In order to continue using the firmware update portion of
this code base, fix up compilation issues in order to run this binary
again. This also disables the kernel driver integration that would allow
this binary to run as a ppm.

BUG=b:383785332
TEST=Compile + run um_ppm -b <bus> -p rts5453 -d on device

Change-Id: I56f7be1fe7a81fe5506221fd09fa4b4f6d5fe0e2