pdc: tps6699x: Add delay prior to sending GET_CABLE_PROPERTY
This CL is a workaround for b:/371053377. If GET_CABLE_PROPERTY is
sent too early relative to USB-PD negotiation with the port partner,
it can lead to subsequent commands (PDC_GET_PDOS) to not
completeing. In addition, this enables GET_SINK_PATH to complete
properly.

This not meant to be a permanent solution, but rather to enable other
functionality for devices using the TI-PDC.

BUG=b:371053377
TEST=Verified that PDC_GET_PDOS doesn't fail when using a cable
matters USBC dock when the port is in a SRC power role. Validated
charging is consitently working with servo_v4p1.
./twister -c -T zephyr/test/pdc --> completes successfully

Change-Id: I6fea926451c0ef3462421a7a3e5635649892fd44