awasuki: modify als count abnormal
Add a bit in the EEPROM to record the normal triggering
of the ALS process. After the ALS is triggered, repeatedly
plug and unplug the AC. When the battery is activated, and
this bit is set. even the ALS process is triggered again,
it will not be counted.

BUG=b:353445852
TEST=Remove the cover, wait for a while, plug in AC,
the count does not increase by two.

Change-Id: I2d06b72009da05a1107bc154f0bf38ca125df021