fpsensor: clarify fp_maintenance() error reporting
This commit clarifies the documentation for the `fp_maintenance()`
function to accurately describe its error reporting behaviour. The
updated documentation explains that the function can update a
`error_state` variable, which is used to store information about detected
dead pixels.

This clarification ensures that users of the function understand how to
better implement it in order to retrieve information about dead pixels
and handle potential errors.

BUG=b:376708279
TEST=CQ

Change-Id: I2d8b30b148b53f827321d52d2f71f50370755699