pdc: Refactor active charge port handling
pdc_power_mgmt attempts to tarck the active charge port by updating a
boolean active_charge field that is a property of each port. Keeping
this value up to date is vulnerable to race conditions.

Instead, charge_manager should be the source of truth regarding the
active charging port. Instead of updating and reading active_charge,
call charge_manager_get_active_charge_port() in order to check if the
current port is supposed to be active.

pdc_power_mgmt_set_active_charge_port() now is only used by charge
manager to notify the PDC subsystem that a change to the active charge
port has been made and to trigger an update by way of raising policy
flags on each of the ports.

BUG=None
TEST=./twister -T zephyr/test/pdc --toolchain {llvm|host}
TEST=Extensive on-device manual tests with two chargers (PD and Type-C),
TEST=battery cutoffs, EC reboots, and ectool chargeoverride.

Change-Id: I3debae59ed66793916dd4060ee563d7061daa0b3