docs: Change fingerprint software write protect verification
Due to a bug in firmware (http://b/375982132), downloading "raw"
fingerprint captures on systems containing an FPC sensor errors
out. This issue has been corrected, but for now it is safe and
potentially advantageous to use the PNM download mode to verify
that software write protect is enabled.

This might be advantageous to use this mode because you could
attempt to parse the beginning header of the download PNM, since
it is just the text "P2".

BUG=b:375982132
TEST=Run "ectool --name=cros_fp fpframe" on buccaneer and
     bloonchipper in both wirte protected and unwriteprotected
     modes.

Change-Id: I3eb025cc1d1316a57470a0ba87e3ee21b31d22c0