pdc: tps6699x: Check for pending interrupts at end of callback
This CL modifies st_irq_run to check the level of the interrupt signal
to the EC prior to exiting. This is required to handle situations
where the PDC sets more than one interrupt status bit in quick
succession.

This situation was happening often following a power role swap request
from a port partner. The status change bit (bit 26) is set initially
and the power_swap_complete bit (bit 4) gets just around the time the
EC is reading the interrupt status register.

BUG=b:373454623
TEST=Verfied on Trulo that power role swaps can be initiated from
servo_v4p1 and that the PDM accurately tracks the attached state and
subsequent connections work.

Change-Id: I970c9b66a7b9e1714b2f91e40957de3d813e4012