board/hyperdebug: Support bit-banging in quick succession
The interrupt handler used for bit-banging is highly optimized, and in
order for the special case of processing the last sample not to slow
down ordinary processing, it is only handled on the next timer interrupt
(after the interrupt that applied the last sample to the GPIO pins).
This saves a few instructions at the end of the handler, which is useful
at clock rates in the hundredths of kilohertz.  However, with very slow
bit banging, e.g. a clock of a tenth of a second, it could result in the
host PC having received acknowledgement that the operation is done, and
turning around and asking for more bit-banging, before the final
interrupt.  This used to result in an error that "another bit-banging
operation is ongoing".

This change makes the non-time-critical USB communication code, which
sends data back to the PC look if it was the final byte, and if so,
disable timer interrupts (this is of course in a race with the next
timer interrupt firing, which would also disable itself, so benign).

BUG=none
TEST=make tast TAST_EXPR=gscdevboard.Ti50RBOX.clamshell  (from ti50/common)

Change-Id: I5e582c8cdd6d65678af2383bf99299978c7ddcc7