board/hyperdebug: ADC calibration robustness
In case of ADC issues (such as improper initialization), the calibration
readings could come back all zeroes, which would previously lead to
division by zero, as the factor_div became zero.

This CL introduces detection if the calibration reading is out of a
reasonable range, and if so, defaults to some hard-coded and "safe" ADC
factors.

ADC and DAC performance will of course not be guaranteed in such cases,
but at least HyperDebug will not panic and drop the USB connection, but
instead send an textual error message to the PC host and console,
allowing something meaningful to surface in test logs.

Also, use clock_enable_module() rather than direct register
manipulation, and no need for explicit ADC reading from within
board_gpio_init(), as that is already performed by calibrate_adc()
(which was introduced later).

BUG=none
TEST=make tast TAST_EXPR=gscdevboard.Ti50RBOX.clamshell  (from ti50/common)

Change-Id: I259dada1ee622dddf02402614d3dd72a2aafe4a4