common: Introduce queue_flush()
Add a new method queue_flush() to be used together with
e.g. queue_add_units(), such that a producer can inform the consumer
when it is OK to leave some bytes/units in the queue waiting for more,
and when the queue needs to be flushed without delay.

This is intended for the usb-stream consumer, which can greatly benefit
from completely filling 64-byte USB packets, rather than
transmitting a single byte at a time, as they arrive on a UART.

Consumers will be notified of a request to flush by getting a call of
consumer_ops.written() with a zero length.

Some consumers will not know about the flusing semantics, but that is
fine, they will ignore a notification with zero length, and they will
eagerly process data, which is always allowed.

Some producers will not make use of queue_flush(), and care has to be
taken that a consumer does not wait infinitely for a flush request that
never comes.  To handle this, consumers must eagerly process data until
the first time they get a notification with zero length, which serves as
indication that the producer uses the flush semantics.  Producers can
make a call to queue_flush() as part of their initialization, to avoid
any actual data being eagerly processed.

BUG=b:378954948
TEST=make tast TAST_EXPR=gscdevboard.GSCSysinfo  (in ti50/common)

Change-Id: I0f6dd4bcb582b164a28f91ce65ef977f78c3198e