chip/stm32: USB stream driver delays outbound USB packet until flush
Adapt code in usb-stream.c to treat zero-length notifications as
requests to flush the queue of outbound data.  This will improve
efficiency of forwarding of UART data to USB endpoints, as it avoids
one-byte USB packets, and instead waits to fill a 64-byte packet, or
until the UART has been idle for a short period.

BUG=b:378954948
TEST=make tast TAST_EXPR=gscdevboard.GSCSysinfo  (in ti50/common)

Change-Id: I55a6b6bfd4728440db52410683e2ee18384f7751